requirements.txt
Django>=5.1,<6
djangorestframework>=3.15,<4
google-auth>=2.35,<3
requests>=2.32,<3
